# doc/a3s-openapi.yaml
openapi: 3.0.0
info:
  description: >-
    API Definition for A3S. This service allows authentication, authorisation and accounting.
  version: 1.1.3
  title: A3S
paths:
...
  '/clients':
    get:
      tags:
        - Client
      description: Search for clients.
      summary: Search for clients.
      operationId: listClients
      parameters:
        - name: page
          in: query
          required: false
          description: The page to view.
          schema:
            type: integer
            minimum: 1
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          description: The size of a page.
          schema:
            type: integer
            minimum: 1
            maximum: 20
            format: int32
            default: 10
        - name: filterName
          in: query
          required: false
          description: A search query filter on the name of the client.
          schema:
            type: string
            minLength: 0
            maxLength: 255
        - name: filterClientId
          in: query
          required: false
          description: A search query filter on the client ID of the client.
          schema:
            type: string
            minLength: 0
            maxLength: 255
          example: FilterByEntityType
        - name: orderBy
          in: query
          required: false
          style: form
          explode: false
          description: >-
            a comma separated list of fields in their sort order. Ascending
            order is assumed. Append '_desc' after a field to indicate descending order.  Supported fields. 'name', 'clientId'
          schema:
            type: string
      responses:
        '200':
          description: OK
          headers:
            X-Pagination:
              description: Pagination header.
              schema:
                $ref: '#/components/schemas/PaginationHeaderResponse'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Oauth2Client'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access the list of clients.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Client list not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/clients/{clientId}':
    get:
      tags:
        - Client
      summary: Get a client.
      description: Get a client by its client Id.
      operationId: getClient
      parameters:
        - name: clientId
          in: path
          description: client
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Oauth2Client'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access the client.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Client not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  #endregion - Clients
  #region - Applications
...
components:
  schemas:
...
    Oauth2Client:
      description: |
        Models an Oauth2 or OpenID Connect client.
      type: object
      required:
        - clientId
        - name
        - allowedGrantTypes
        - allowedScopes
        - allowedOfflineAccess
      properties:
        clientId:
          type: string
          description: The unique ID of the client.
        name:
          type: string
          description: Client display name. Used for logging and the consent screen.
        allowedGrantTypes:
          type: array
          items:
            type: string
          description: The Oauth2 grant types that the client will be permitted to use.
        redirectUris:
          description: Specifies the allowed URIs to return tokens or authorisation codes to.
          type: array
          items:
            type: string
            format: url
        allowedCorsOrigins:
          description: Sets the allowed CORS origins for JavaScript clients.
          type: array
          items:
            type: string
            format: url
        postLogoutRedirectUris:
          description: Specifies the allowed URIs to redirect to after logout.
          type: array
          items:
            type: string
            format: url
        allowedScopes:
          description: Specifies the scopes that the client is allowed to access. If empty, the client cannot access any scopes.
          type: array
          items:
            type: string
        allowedOfflineAccess:
          description: Defines whether offline access with refresh tokens is permitted for this client.
          type: boolean
    # End Region - Clients
    
    # Begin Region - Application Security Definition
...
    ErrorResponse:
      description: Error response object to give more info with HTTP status code
      type: object
      required:
        - message
      properties:
        message:
          type: string
    PaginationHeaderResponse:
      description: >-
        This is the standard response for requests that have paginated
        collections
      type: object
      properties:
        total:
          type: integer
          format: int32
          minimum: 1
          description: The total number of pages in the result set
        count:
          type: integer
          format: int32
          minimum: 1
          description: The total number of results in the result set
        size:
          type: integer
          format: int32
          minimum: 1
          maximum: 20
          default: 10
          description: >-
            The number items to include in a page of results. The page size for
            the result set
        current:
          type: integer
          format: int32
          minimum: 1
          description: >-
            The position of the page in the paged result set that is being
            returned
        first:
          type: string
          format: url
          description: >-
            The link to the first page of results containing [size] results.
            This link includes the page, size, filter, orderBy and fields query
            parameters.
        last:
          type: string
          format: url
          description: >-
            The link to the last page of results containing [size] results. This
            link includes the page, size, filter, orderBy and fields query
            parameters.
        prev:
          type: string
          format: url
          description: >-
            The link to the previous page of results containing [size] results.
            This link includes the page, size, filter, orderBy and fields query
            parameters. If this is the first page in the result set then this
            will be [null].
        next:
          type: string
          format: url
          description: >-
            The link to the next page of results containing [size] results. This
            link includes the page, size, filter, orderBy and fields query
            parameters. If this is the last page in the result set then this will
            be [null].
...